# [0.0.3] 'To-day not To-week'
It may look like there were big changes because my code base is tiny.

The back end was a lot quicker to finish than last time, but I suppose it was because I was lost before and now I am a little bit less lost when it comes to sqlite3.

Anyway, it is time for me to integrate the front end and back end to a much greater degree - i.e., to use both and to collect data for the eventual data science with pandas.

### Added
-[Backend] Tests to the python in case there is no JSON object in the call.
-[Backend] Guard clauses for ease of use.
-[Backend] Much more functionalilty, e.g., SELECT, INSERT, DELETE, etc.

### Changed
-[Backend] The SQL now only accepts entries by the day rather than by the week: i.e., no week_id.
-[Backend] The python only sets one value, and that is the boolean isUsed.
-[Backend] The python now uses JSON rather than a dot ini file.
-[Backend] Alterations to repr that allow for automatic new lines and '>' before the input string for ease of reading the console log.

-Changed naming scheme to toggle_tracker from toggle-tracker.
-Changed name of CSS and pyton file.

### Removed
-All unimportant files that are no longer used: configurations.py, virtue-text.py.
-[Backend] Week-based functions are to be removed when I finish implementing the day-based ones.