Major Rework of Structure, in the code base and in the file system
Some of these items are still from documentation, but I will edit them out later.
I wanted the base to be accomplished and to force myself to work like a web developer and software engineer rather than write code in random places and expect it all to work – burning out once it doesn't. It was not the best of times.