v 0.1.0 the barest of bones
I am making a design doc, though I have not finished it before I did this. I will for the next few versions.

Right now, I have not implemented most of the code for the MVP, but I have made a few more pages and architectural changes.

There is yet no database or model, and I require a different view engine to incorporate block inheritance.

I have incorporated a way for JSON to travel to and from the server and browser, but I still authentication, authorisation, validation, models, and a database to save the data.